moved themes to separate config files under theme/, needs polish
