+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/DarcerulaThemeConfig.java
@@ -1,0 +1,7 @@
+package org.quiltmc.enigma.gui.config.theme;
+
+public class DarcerulaThemeConfig extends ThemeCreator {
+	public DarcerulaThemeConfig() {
+		super(ThemeProperties.DARCERULA);
+	}
+}
