+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/DarculaThemeConfig.java
@@ -1,0 +1,7 @@
+package org.quiltmc.enigma.gui.config.theme;
+
+public class DarculaThemeConfig extends ThemeCreator {
+	public DarculaThemeConfig() {
+		super(ThemeProperties.DARCULA);
+	}
+}
