+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/DefaultThemeConfig.java
@@ -1,0 +1,7 @@
+package org.quiltmc.enigma.gui.config.theme;
+
+public class DefaultThemeConfig extends ThemeCreator {
+	public DefaultThemeConfig() {
+		super(ThemeProperties.DEFAULT);
+	}
+}
