+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/MetalThemeConfig.java
@@ -1,0 +1,7 @@
+package org.quiltmc.enigma.gui.config.theme;
+
+public class MetalThemeConfig extends ThemeCreator {
+	public MetalThemeConfig() {
+		super(ThemeProperties.METAL);
+	}
+}
