+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/NoneThemeConfig.java
@@ -1,0 +1,7 @@
+package org.quiltmc.enigma.gui.config.theme;
+
+public class NoneThemeConfig extends ThemeCreator {
+	public NoneThemeConfig() {
+		super(ThemeProperties.NONE);
+	}
+}
