+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/SystemThemeConfig.java
@@ -1,0 +1,7 @@
+package org.quiltmc.enigma.gui.config.theme;
+
+public class SystemThemeConfig extends ThemeCreator {
+	public SystemThemeConfig() {
+		super(ThemeProperties.SYSTEM);
+	}
+}
