+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/look_and_feel/ConfigurableFlatLightLaf.java
@@ -1,12 +1,12 @@
 package org.quiltmc.enigma.gui.config.theme.look_and_feel;
 
 import com.formdev.flatlaf.FlatLightLaf;
-import org.quiltmc.enigma.gui.config.theme.Theme;
+import org.quiltmc.enigma.gui.config.theme.ThemeCreator;
 
 public class ConfigurableFlatLightLaf extends ConfigurableFlatLaf {
 	public static final String NAME = "Configurable " + FlatLightLaf.NAME;
 
-	public ConfigurableFlatLightLaf(Theme.LookAndFeelColors colors) {
+	public ConfigurableFlatLightLaf(ThemeCreator.LookAndFeelColors colors) {
 		super(NAME, colors, Base.LIGHT);
 	}
 }
