+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/look_and_feel/SerializableColorGetter.java
@@ -1,17 +1,17 @@
 package org.quiltmc.enigma.gui.config.theme.look_and_feel;
 
 import org.quiltmc.config.api.values.TrackedValue;
-import org.quiltmc.enigma.gui.config.theme.Theme;
+import org.quiltmc.enigma.gui.config.theme.ThemeCreator;
 
 import java.util.function.Function;
 
 @FunctionalInterface
-public interface SerializableColorGetter extends Function<Theme.LookAndFeelColors, Theme.SerializableColor> {
+public interface SerializableColorGetter extends Function<ThemeCreator.LookAndFeelColors, ThemeCreator.SerializableColor> {
 	static SerializableColorGetter of(TrackedSerializableColorGetter getter) {
 		return of(getter.andThen(TrackedValue::value));
 	}
 
-	static SerializableColorGetter of(Function<Theme.LookAndFeelColors, Theme.SerializableColor> getter) {
+	static SerializableColorGetter of(Function<ThemeCreator.LookAndFeelColors, ThemeCreator.SerializableColor> getter) {
 		return getter::apply;
 	}
 }
