+++ enigma-swing/src/main/java/org/quiltmc/enigma/gui/config/theme/look_and_feel/TrackedSerializableColorGetter.java
@@ -1,10 +1,10 @@
 package org.quiltmc.enigma.gui.config.theme.look_and_feel;
 
 import org.quiltmc.config.api.values.TrackedValue;
-import org.quiltmc.enigma.gui.config.theme.Theme;
+import org.quiltmc.enigma.gui.config.theme.ThemeCreator;
 
 import java.util.function.Function;
 
 @FunctionalInterface
 public interface TrackedSerializableColorGetter extends Function
-		<Theme.LookAndFeelColors, TrackedValue<Theme.SerializableColor>> { }
+		<ThemeCreator.LookAndFeelColors, TrackedValue<ThemeCreator.SerializableColor>> { }
